Correctly delete files after archive is finished
Previous code only deleted correctly the siva file. Some files from the
temporary repository were left behind plus the directory for the job.

Now the temporary dir has this organization:

    <hash>_<time>/{siva,tmp}

For example:

    5d7303c49ac984a9fec60523f2d5297682e16646_1512151128438838822/siva
    5d7303c49ac984a9fec60523f2d5297682e16646_1512151128438838822/tmp

On cleanup the directory is deleted. This cleans siva file and any file
lingering in the tmp repo.