repository: decouple fs logic from copier
Closes #58

This commit introduces the decouple of fs logic from the copiers, that become a single copier.

Before we had several copiers, one per remote file system. This was problematic because some high-level logic could not be reused and forced us to write again and again the same logic.

Now, we have two pieces that together act as the old copiers.

- Filesystems: which are `hdfsFs` and `localFs` that fulfill a certain interface that `Copier` needs to copy from a fs to the other.
- Copier: which is no longer an interface, but a struct, that knows how to apply this high level logic of copying from fs A to fs B in the exact same way by using the contract defined by the `Fs` interface.

In addition to that, `CopyFromRemote` and `CopyToRemote` are cancellable now.

Signed-off-by: Miguel Molina <miguel@erizocosmi.co>